+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `required` prop to the `Checkbox` component
 - Added `required` prop to the `RadioGroup` component
 - Added `required` prop to the `Radio` component (not visible in the UI)
+- Added `disabled` property to the `Autocomplete` component
 
 ### Fixed